Three scripts to explore signal erp similarity.
An additional three Matlab scripts to compare the envelope of the Grand Average temporal sequence (ERP) extracted for slope analysis with the temporal sequence of individual subject ERPs, with the aim of assessing the robustness of the effect across subjects. The similarity between the Grand Average ERP and individual subjects' ERPs is measured using Spearman's correlation coefficient and Dynamic Time Warping (DTW) similarity. This difference between the GA ERP and the subject-level ERPs is presented as a bar graph presenting the DTW dis-similarity plotted against spearman's correlation coefficient. The user can interact with this plot to view, in detail, a comparison of the GA-ERP and the subject-level ERP.